core: move Runtime tests to runtime package
