+++ pkg/core/interop_system_neotest_test.go
@@ -3,7 +3,6 @@ package core_test
 import (
 	"encoding/json"
 	"fmt"
-	"math"
 	"math/big"
 	"strings"
 	"testing"
@@ -11,12 +10,10 @@ import (
 	"github.com/nspcc-dev/neo-go/internal/contracts"
 	"github.com/nspcc-dev/neo-go/pkg/compiler"
 	"github.com/nspcc-dev/neo-go/pkg/config"
-	"github.com/nspcc-dev/neo-go/pkg/core/interop"
 	"github.com/nspcc-dev/neo-go/pkg/core/interop/interopnames"
 	"github.com/nspcc-dev/neo-go/pkg/core/native/nativenames"
 	"github.com/nspcc-dev/neo-go/pkg/core/transaction"
 	"github.com/nspcc-dev/neo-go/pkg/crypto/keys"
-	"github.com/nspcc-dev/neo-go/pkg/encoding/address"
 	"github.com/nspcc-dev/neo-go/pkg/io"
 	"github.com/nspcc-dev/neo-go/pkg/neotest"
 	"github.com/nspcc-dev/neo-go/pkg/neotest/chain"
@@ -27,56 +24,6 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/vm/stackitem"
 	"github.com/stretchr/testify/require"
 )
-
-func TestSystemRuntimeGetRandom_DifferentTransactions(t *testing.T) {
-	bc, acc := chain.NewSingle(t)
-	e := neotest.NewExecutor(t, bc, acc, acc)
-
-	w := io.NewBufBinWriter()
-	emit.Syscall(w.BinWriter, interopnames.SystemRuntimeGetRandom)
-	require.NoError(t, w.Err)
-	script := w.Bytes()
-
-	tx1 := e.PrepareInvocation(t, script, []neotest.Signer{e.Validator}, bc.BlockHeight()+1)
-	tx2 := e.PrepareInvocation(t, script, []neotest.Signer{e.Validator}, bc.BlockHeight()+1)
-	e.AddNewBlock(t, tx1, tx2)
-	e.CheckHalt(t, tx1.Hash())
-	e.CheckHalt(t, tx2.Hash())
-
-	res1 := e.GetTxExecResult(t, tx1.Hash())
-	res2 := e.GetTxExecResult(t, tx2.Hash())
-
-	r1, err := res1.Stack[0].TryInteger()
-	require.NoError(t, err)
-	r2, err := res2.Stack[0].TryInteger()
-	require.NoError(t, err)
-	require.NotEqual(t, r1, r2)
-}
-
-func TestSystemRuntimeGasLeft(t *testing.T) {
-	const runtimeGasLeftPrice = 1 << 4
-
-	bc, acc := chain.NewSingle(t)
-	e := neotest.NewExecutor(t, bc, acc, acc)
-	w := io.NewBufBinWriter()
-
-	gasLimit := 1100
-	emit.Syscall(w.BinWriter, interopnames.SystemRuntimeGasLeft)
-	emit.Syscall(w.BinWriter, interopnames.SystemRuntimeGasLeft)
-	require.NoError(t, w.Err)
-	tx := transaction.New(w.Bytes(), int64(gasLimit))
-	tx.Nonce = neotest.Nonce()
-	tx.ValidUntilBlock = e.Chain.BlockHeight() + 1
-	e.SignTx(t, tx, int64(gasLimit), acc)
-	e.AddNewBlock(t, tx)
-	e.CheckHalt(t, tx.Hash())
-	res := e.GetTxExecResult(t, tx.Hash())
-	l1 := res.Stack[0].Value().(*big.Int)
-	l2 := res.Stack[1].Value().(*big.Int)
-
-	require.Equal(t, int64(gasLimit-runtimeGasLeftPrice*interop.DefaultBaseExecFee), l1.Int64())
-	require.Equal(t, int64(gasLimit-2*runtimeGasLeftPrice*interop.DefaultBaseExecFee), l2.Int64())
-}
 
 func TestLoadToken(t *testing.T) {
 	bc, acc := chain.NewSingle(t)
@@ -102,63 +49,6 @@ func TestLoadToken(t *testing.T) {
 	})
 	t.Run("invalid contract", func(t *testing.T) {
 		cInvoker.InvokeFail(t, "token contract 0000000000000000000000000000000000000000 not found: key not found", "callT1")
-	})
-}
-
-func TestSystemRuntimeGetNetwork(t *testing.T) {
-	bc, acc := chain.NewSingle(t)
-	e := neotest.NewExecutor(t, bc, acc, acc)
-	w := io.NewBufBinWriter()
-
-	emit.Syscall(w.BinWriter, interopnames.SystemRuntimeGetNetwork)
-	require.NoError(t, w.Err)
-	e.InvokeScriptCheckHALT(t, w.Bytes(), []neotest.Signer{acc}, stackitem.NewBigInteger(big.NewInt(int64(bc.GetConfig().Magic))))
-}
-
-func TestSystemRuntimeGetAddressVersion(t *testing.T) {
-	bc, acc := chain.NewSingle(t)
-	e := neotest.NewExecutor(t, bc, acc, acc)
-	w := io.NewBufBinWriter()
-
-	emit.Syscall(w.BinWriter, interopnames.SystemRuntimeGetAddressVersion)
-	require.NoError(t, w.Err)
-	e.InvokeScriptCheckHALT(t, w.Bytes(), []neotest.Signer{acc}, stackitem.NewBigInteger(big.NewInt(int64(address.NEO3Prefix))))
-}
-
-func TestSystemRuntimeBurnGas(t *testing.T) {
-	bc, acc := chain.NewSingle(t)
-	e := neotest.NewExecutor(t, bc, acc, acc)
-	managementInvoker := e.ValidatorInvoker(e.NativeHash(t, nativenames.Management))
-
-	cs, _ := contracts.GetTestContractState(t, pathToInternalContracts, 0, 1, acc.ScriptHash())
-	rawManifest, err := json.Marshal(cs.Manifest)
-	require.NoError(t, err)
-	rawNef, err := cs.NEF.Bytes()
-	require.NoError(t, err)
-	tx := managementInvoker.PrepareInvoke(t, "deploy", rawNef, rawManifest)
-	e.AddNewBlock(t, tx)
-	e.CheckHalt(t, tx.Hash())
-	cInvoker := e.ValidatorInvoker(cs.Hash)
-
-	t.Run("good", func(t *testing.T) {
-		h := cInvoker.Invoke(t, stackitem.Null{}, "burnGas", int64(1))
-		res := e.GetTxExecResult(t, h)
-
-		t.Run("gas limit exceeded", func(t *testing.T) {
-			tx := e.NewUnsignedTx(t, cs.Hash, "burnGas", int64(2))
-			e.SignTx(t, tx, res.GasConsumed, acc)
-			e.AddNewBlock(t, tx)
-			e.CheckFault(t, tx.Hash(), "GAS limit exceeded")
-		})
-	})
-	t.Run("too big integer", func(t *testing.T) {
-		gas := big.NewInt(math.MaxInt64)
-		gas.Add(gas, big.NewInt(1))
-
-		cInvoker.InvokeFail(t, "invalid GAS value", "burnGas", gas)
-	})
-	t.Run("zero GAS", func(t *testing.T) {
-		cInvoker.InvokeFail(t, "GAS must be positive", "burnGas", int64(0))
 	})
 }
